Eliminar repositorios ATMConfigurationRepository
...y CashManagementRepository

+++ ATM.Machine/ATM.Machine/ATM.Application/Interfaces/Services/IATMConfigurationService.cs
@@ -4,7 +4,7 @@ namespace ATM.Application.Interfaces.Services
 {
     public interface IATMConfigurationService
     {
-        Task SetWithdrawalLimitsAsync(int maxLimit, int minLimit);
-        Task<(int MaxLimit, int MinLimit)> GetWithdrawalLimitsAsync();
+        Task SetWithdrawalLimitsAsync(decimal maxLimit, decimal minLimit);
+        Task<(decimal MaxLimit, decimal MinLimit)> GetWithdrawalLimitsAsync();
     }
 }
